requirements.txt
Flask
Blueprint
flask_sqlalchemy
mysql-connector
pymysql
mysqlclient
